requirements.txt
django
requests
transformers
openai
google-auth-oauthlib
google-api-python-client
scikit-learn
twilio